mailing/pyproject.toml
[tool.poetry]
name = "mailing"
version = "1.0.0"
description = "Simple email service with RabbitMQ and Celery"
authors = ["ME <pvs.versia@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
pydantic = {extras = ["email"], version = "^2.10.5"}
pydantic-settings = "^2.6.1"
celery = "^5.3.6"
pika = "^1.3.2"
jinja2 = "^3.1.3"
python-dotenv = "^1.0.0"
cryptography = "^42.0.0"

[tool.poetry.group.test.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.21.1"
pytest-cov = "^4.1.0"
pytest-mock = "^3.12.0"
httpx = "^0.26.0"
requests = "^2.31.0"

[tool.poetry.group.dev.dependencies]
black = "^24.10.0"
isort = "^5.13.2"
flake8 = "^7.1.1"
ruff = "^0.9.2"

[[tool.poetry.source]]
name = "tsinghua"
url = "https://pypi.tuna.tsinghua.edu.cn/simple/"
priority = "primary"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
